requirements.txt
streamlit>=1.28.0
opencv-python-headless==4.8.1.78
torch>=2.5.0
torchvision>=0.20.0
Pillow>=10.0.0
numpy>=1.24.0
ultralytics>=8.0.0
folium>=0.15.1
plotly>=5.18.0
requests>=2.31.0 